Rocket-Launch/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile
#
aiohttp==3.9.3
    # via
    #   datasets
    #   fsspec
aiosignal==1.3.1
    # via aiohttp
async-timeout==4.0.3
    # via aiohttp
attrs==23.2.0
    # via aiohttp
certifi==2024.2.2
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via nltk
cmake==3.28.1
    # via triton
contourpy==1.2.0
    # via matplotlib
cycler==0.12.1
    # via matplotlib
datasets==2.15.0
    # via -r requirements.in
dill==0.3.7
    # via
    #   datasets
    #   multiprocess
filelock==3.13.1
    # via
    #   huggingface-hub
    #   torch
    #   transformers
    #   triton
fonttools==4.48.1
    # via matplotlib
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
fsspec[http]==2023.10.0
    # via
    #   datasets
    #   fsspec
    #   huggingface-hub
    #   pytorch-lightning
huggingface-hub==0.20.3
    # via
    #   datasets
    #   tokenizers
    #   transformers
idna==3.6
    # via
    #   requests
    #   yarl
importlib-resources==6.1.1
    # via matplotlib
jinja2==3.1.3
    # via torch
joblib==1.3.2
    # via nltk
kiwisolver==1.4.5
    # via matplotlib
lightning-utilities==0.10.1
    # via
    #   pytorch-lightning
    #   torchmetrics
lit==17.0.6
    # via triton
markupsafe==2.1.5
    # via jinja2
matplotlib==3.8.1
    # via -r requirements.in
mpmath==1.3.0
    # via sympy
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
multiprocess==0.70.15
    # via datasets
networkx==3.2.1
    # via torch
nltk==3.8.1
    # via -r requirements.in
numpy==1.26.4
    # via
    #   contourpy
    #   datasets
    #   matplotlib
    #   pandas
    #   pyarrow
    #   pytorch-lightning
    #   torchmetrics
    #   transformers
nvidia-cublas-cu11==11.10.3.66
    # via
    #   nvidia-cudnn-cu11
    #   nvidia-cusolver-cu11
    #   torch
nvidia-cuda-cupti-cu11==11.7.101
    # via torch
nvidia-cuda-nvrtc-cu11==11.7.99
    # via torch
nvidia-cuda-runtime-cu11==11.7.99
    # via torch
nvidia-cudnn-cu11==8.5.0.96
    # via torch
nvidia-cufft-cu11==10.9.0.58
    # via torch
nvidia-curand-cu11==10.2.10.91
    # via torch
nvidia-cusolver-cu11==11.4.0.1
    # via torch
nvidia-cusparse-cu11==11.7.4.91
    # via torch
nvidia-nccl-cu11==2.14.3
    # via torch
nvidia-nvtx-cu11==11.7.91
    # via torch
packaging==23.2
    # via
    #   datasets
    #   huggingface-hub
    #   lightning-utilities
    #   matplotlib
    #   pytorch-lightning
    #   torchmetrics
    #   transformers
pandas==2.2.0
    # via
    #   -r requirements.in
    #   datasets
pillow==10.2.0
    # via matplotlib
psutil==5.9.6
    # via -r requirements.in
pyarrow==15.0.0
    # via datasets
pyarrow-hotfix==0.6
    # via datasets
pyparsing==3.1.1
    # via matplotlib
python-dateutil==2.8.2
    # via
    #   matplotlib
    #   pandas
pytorch-lightning==2.1.2
    # via -r requirements.in
pytz==2024.1
    # via pandas
pyyaml==6.0.1
    # via
    #   -r requirements.in
    #   datasets
    #   huggingface-hub
    #   pytorch-lightning
    #   transformers
regex==2023.12.25
    # via
    #   nltk
    #   transformers
requests==2.31.0
    # via
    #   datasets
    #   fsspec
    #   huggingface-hub
    #   transformers
safetensors==0.4.2
    # via transformers
sentencepiece==0.1.99
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
sympy==1.12
    # via torch
tokenizers==0.15.1
    # via transformers
torch==2.0.1
    # via
    #   -r requirements.in
    #   pytorch-lightning
    #   torchmetrics
    #   triton
torchmetrics==1.3.0.post0
    # via pytorch-lightning
tqdm==4.66.1
    # via
    #   -r requirements.in
    #   datasets
    #   huggingface-hub
    #   nltk
    #   pytorch-lightning
    #   transformers
transformers==4.35.2
    # via -r requirements.in
triton==2.0.0
    # via torch
typing-extensions==4.9.0
    # via
    #   huggingface-hub
    #   lightning-utilities
    #   pytorch-lightning
    #   torch
tzdata==2023.4
    # via pandas
urllib3==2.2.0
    # via requests
wheel==0.42.0
    # via
    #   nvidia-cublas-cu11
    #   nvidia-cuda-cupti-cu11
    #   nvidia-cuda-runtime-cu11
    #   nvidia-curand-cu11
    #   nvidia-cusparse-cu11
    #   nvidia-nvtx-cu11
xxhash==3.4.1
    # via datasets
yarl==1.9.4
    # via aiohttp
zipp==3.17.0
    # via importlib-resources

# The following packages are considered to be unsafe in a requirements file:
# setuptools
